pustak/requirements.txt
aioredis==1.3.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-allauth==0.51.0
django-appconf==1.0.5
django-avatar==7.0.0
django-bootstrap-form==3.4
django-cleanup==6.0.0
dnspython==2.2.1
email-validator==1.3.0
gunicorn
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.7
httpx==0.18.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pyuploadcare==3.0.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.4.0
